+++ src/parse_data.c
@@ -6,7 +6,7 @@
 /*   By: jdomingu <jdomingu@student.42malaga.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/22 21:09:35 by jdomingu          #+#    #+#             */
-/*   Updated: 2022/12/22 21:12:12 by jdomingu         ###   ########.fr       */
+/*   Updated: 2023/02/20 21:48:12 by jdomingu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -79,7 +79,7 @@ static void	ft_get_numbers_of_string(t_stack *a, int *b, char *numbers)
 void	check_range_and_duplicates(int ac, char **av, t_stack *a, t_stack *b)
 {
 	int	i;
-	int	parse_to_int;
+	long	parse_to_int;
 
 	i = 0;
 	while (++i < ac)
